# Optimize Floki.children