Floki.children is called a lot by Phoenix LiveView when doing tests. On the project I'm working on, this functions gets called more than 8 million times when running all of our tests. 

This PR adds some pattern matching to this function to avoid using `Access.get` if possible.

```
Name                      ips        average  deviation         median         99th %
children (pr)         11.50 K       86.95 μs   ±239.29%       53.26 μs     1559.21 μs
children (main)        7.19 K      139.07 μs   ±145.06%      104.25 μs     1507.63 μs

Comparison: 
children (pr)         11.50 K
children (main)        7.19 K - 1.60x slower +52.12 μs
``` 
```

read_file = fn name ->
  __ENV__.file
  |> Path.dirname()
  |> Path.join(name)
  |> File.read!()
  |> Floki.parse_document!()
end

doc = read_file.("small.html")
html_nodes = Floki.find(doc, "*")

Benchee.run(
  %{
    "children" => fn ->
      for html_node <- html_nodes do
        Floki.children(html_node)
        Floki.children(html_node, include_text: false)
        Floki.children(html_node, include_text: true)
      end
    end
  },
  time: 10,
  save: [path: "benchs/results/finder-#{tag}.benchee", tag: tag],
  memory_time: 2
)
```